library/axi_pwm_custom: Completed and simulated the pwm module
Completed the PWM module by editing this 2 files:
- axi_pwm_custom.v: added the connection samples for every data_channel, in order to make sure of the connection between the larger module of the PWM and the input
from the ADC
- axi_pwm_custom_if: added the whole logic of the pwm signal. The signal is generated using a counter with an end_of period signal, in order to control the fact that
the data from the adc should be processed only after the end of the pwm period. The pwm will be visible on the leds of the fpga, which are connected at the end of
this file. The turning on and off logic can be seen at the always loop, where all 6 pwm signals are controlled using the values of the count.

The simulation of the PWM can be seen in the photo attached, PWM_Cosmin. The simulation was made using Vivado and the signals can be seen at the end, 6 signals which
are high for shorter and increasingly higher, and after saturation, it restarts as zero and re-enters the loop

Signed-off-by: Cosmin Popa <popa.gi.cosmin@student.utcluj.ro>